change it so template type and not int

+++ Application/Templates/Mapping/TemplateMapping.cs
@@ -2,6 +2,7 @@
 using Application.Templates.GetById;
 using Domain.Templates;
 using Domain.Templates.Fitness;
+using SharedKernel;
 using SharedKernel.Responses;
 
 
@@ -29,7 +30,7 @@ namespace Application.Templates.Mapping
                     UserId = workout.UserId,
                     Name = workout.Name,
                     Description = workout.Description,
-                    TemplateType = (int)TemplateType.Workout,
+                    TemplateType = TemplateType.Workout,
                     Activities = workoutActivities != null && workoutActivities.TryGetValue(workout.Id, out var activities)
                         ? activities.Select(MapWorkoutActivity).ToList()
                         : workout.Activities?.Select(MapWorkoutActivity).ToList() ?? new List<WorkoutActivityResponse>()
@@ -47,7 +48,7 @@ namespace Application.Templates.Mapping
                         UserId = fitness.UserId,
                         Name = fitness.Name,
                         Description = fitness.Description,
-                        TemplateType = (int)TemplateType.Fitness,
+                        TemplateType = TemplateType.Fitness,
                         Activities = fitnessActivities != null && fitnessActivities.TryGetValue(fitness.Id, out var activities)
                             ? activities.Select(MapFitnessActivity).ToList()
                             : fitness.Activities?.Select(MapFitnessActivity).ToList() ?? new List<FitnessActivityResponse>()
